Move `BufferPool` to `internal` package of `core` module because of its exposure in metadata API
