+++ rsocket-core/src/commonMain/kotlin/io/rsocket/kotlin/Connection.kt
@@ -19,7 +19,6 @@ package io.rsocket.kotlin
 import io.ktor.utils.io.core.*
 import io.rsocket.kotlin.frame.*
 import io.rsocket.kotlin.internal.*
-import io.rsocket.kotlin.internal.io.*
 import kotlinx.coroutines.*
 
 /**
