+++ rsocket-core/src/commonMain/kotlin/io/rsocket/kotlin/core/RSocketConnectorBuilder.kt
@@ -18,7 +18,6 @@ package io.rsocket.kotlin.core
 
 import io.rsocket.kotlin.*
 import io.rsocket.kotlin.internal.*
-import io.rsocket.kotlin.internal.io.*
 import io.rsocket.kotlin.keepalive.*
 import io.rsocket.kotlin.logging.*
 import io.rsocket.kotlin.payload.*
