+++ rsocket-core/src/commonMain/kotlin/io/rsocket/kotlin/frame/ExtensionFrame.kt
@@ -18,7 +18,7 @@ package io.rsocket.kotlin.frame
 
 import io.ktor.utils.io.core.*
 import io.rsocket.kotlin.frame.io.*
-import io.rsocket.kotlin.internal.io.*
+import io.rsocket.kotlin.internal.*
 import io.rsocket.kotlin.payload.*
 
 internal class ExtensionFrame(
