+++ rsocket-core/src/commonMain/kotlin/io/rsocket/kotlin/frame/ResumeFrame.kt
@@ -18,7 +18,7 @@ package io.rsocket.kotlin.frame
 
 import io.ktor.utils.io.core.*
 import io.rsocket.kotlin.frame.io.*
-import io.rsocket.kotlin.internal.io.*
+import io.rsocket.kotlin.internal.*
 
 internal class ResumeFrame(
     val version: Version,
