+++ rsocket-core/src/commonMain/kotlin/io/rsocket/kotlin/internal/Connect.kt
@@ -19,7 +19,6 @@ package io.rsocket.kotlin.internal
 import io.rsocket.kotlin.*
 import io.rsocket.kotlin.core.*
 import io.rsocket.kotlin.frame.*
-import io.rsocket.kotlin.internal.io.*
 import kotlinx.coroutines.*
 
 @OptIn(TransportApi::class)
