+++ rsocket-core/src/commonMain/kotlin/io/rsocket/kotlin/internal/FrameSender.kt
@@ -19,7 +19,6 @@ package io.rsocket.kotlin.internal
 import io.ktor.utils.io.core.*
 import io.rsocket.kotlin.frame.*
 import io.rsocket.kotlin.frame.io.*
-import io.rsocket.kotlin.internal.io.*
 import io.rsocket.kotlin.payload.*
 import kotlinx.coroutines.*
 import kotlin.math.*
