+++ rsocket-core/src/commonMain/kotlin/io/rsocket/kotlin/metadata/Metadata.kt
@@ -19,7 +19,7 @@ package io.rsocket.kotlin.metadata
 import io.ktor.utils.io.core.*
 import io.rsocket.kotlin.*
 import io.rsocket.kotlin.core.*
-import io.rsocket.kotlin.internal.io.*
+import io.rsocket.kotlin.internal.*
 import io.rsocket.kotlin.payload.*
 
 @ExperimentalMetadataApi
