+++ rsocket-core/src/commonMain/kotlin/io/rsocket/kotlin/metadata/security/BearerAuthMetadata.kt
@@ -18,7 +18,7 @@ package io.rsocket.kotlin.metadata.security
 
 import io.ktor.utils.io.core.*
 import io.rsocket.kotlin.*
-import io.rsocket.kotlin.internal.io.*
+import io.rsocket.kotlin.internal.*
 
 @ExperimentalMetadataApi
 public class BearerAuthMetadata(
