+++ rsocket-core/src/commonMain/kotlin/io/rsocket/kotlin/metadata/security/RawAuthMetadata.kt
@@ -19,7 +19,7 @@ package io.rsocket.kotlin.metadata.security
 import io.ktor.utils.io.core.*
 import io.rsocket.kotlin.*
 import io.rsocket.kotlin.frame.io.*
-import io.rsocket.kotlin.internal.io.*
+import io.rsocket.kotlin.internal.*
 
 @ExperimentalMetadataApi
 public class RawAuthMetadata(
